+++ src/ocrsmith/core/backgrounds/strategies/__init__.py
@@ -1,4 +1,0 @@
-from .GradientBackground import GradientBackground
-from .ImageBackground import ImageBackground
-from .TextureBackground import TextureBackground
-from .SolidColorBackground import SolidColorBackground