+++ src/ocrsmith/core/text_placement/strategies/__init__.py
@@ -1,5 +1,0 @@
-from .CenterPlacementStrategy import CenterPlacementStrategy
-from .RandomPlacementStrategy import RandomPlacementStrategy
-from .GridPlacementStrategy import GridPlacementStrategy
-from .PageNumberPlacementStrategy import PageNumberPlacementStrategy
-from .PageTitlePlacementStrategy import PageTitlePlacementStrategy