+++ src/ocrsmith/datasets/TextLoaderInterface.py
@@ -2,12 +2,12 @@
 
 from abc import ABC, abstractmethod
 
+
 class TextLoaderInterface(ABC):
     @abstractmethod
     def load_texts():
         pass
-    
+
     @abstractmethod
     def __iter__(self):
         pass
-