+++ src/ocrsmith/datasets/__init__.py
@@ -1,2 +1,15 @@
+"""Text sources in, datasets out."""
+
+from .loaders import CSVTextLoader, HuggingFaceTextLoader, ParquetTextLoader
 from .TextDataManager import TextDataManager
-from .loaders import CSVTextLoader, ParquetTextLoader, HuggingFaceTextLoader
+from .writers import SampleSink, build_sink, sink_names
+
+__all__ = [
+    "CSVTextLoader",
+    "HuggingFaceTextLoader",
+    "ParquetTextLoader",
+    "SampleSink",
+    "TextDataManager",
+    "build_sink",
+    "sink_names",
+]